+++ src/helpers/const.js
@@ -1,0 +1,2 @@
+export const DEFAULT_SIGNATURE_LENGTH = 8192;
+export const DEFAULT_BYTE_RANGE_PLACEHOLDER = '**********';
