+++ src/helpers/index.js
@@ -1,0 +1,4 @@
+export {default as extractSignature} from './extractSignature';
+export {default as pdfkitAddPlaceholder} from './pdfkitAddPlaceholder';
+export {default as plainAddPlaceholder} from './plainAddPlaceholder';
+export {default as removeTrailingNewLine} from './removeTrailingNewLine';
