+++ src/helpers/plainAddPlaceholder.js
@@ -1,34 +1,9 @@
 import fs from 'fs';
-import {DEFAULT_BYTE_RANGE_PLACEHOLDER} from './signpdf';
-import SignPdfError from './SignPdfError';
-import PDFObject, {PDFReferenceMock} from './pdfobject';
-
-export const DEFAULT_SIGNATURE_LENGTH = 8192;
-
-/**
- * Removes a trailing new line if there is such.
- *
- * Also makes sure the file ends with an EOF line as per spec.
- * @param {Buffer} pdf
- * @returns {Buffer}
- */
-export const removeTrailingNewLine = (pdf) => {
-    const lastChar = pdf.slice(pdf.length - 1).toString();
-    if (lastChar === '\n') {
-        // remove the trailing new line
-        return pdf.slice(0, pdf.length - 1);
-    }
-
-    const lastLine = pdf.slice(pdf.length - 6).toString();
-    if (lastLine !== '\n%%EOF') {
-        throw new SignPdfError(
-            'A PDF file must end with an EOF line.',
-            SignPdfError.TYPE_PARSE,
-        );
-    }
-
-    return pdf;
-};
+import SignPdfError from '../SignPdfError';
+import PDFObject, {PDFReferenceMock} from '../pdfobject';
+import removeTrailingNewLine from './removeTrailingNewLine';
+import {DEFAULT_SIGNATURE_LENGTH} from './const';
+import pdfkitAddPlaceholder from './pdfkitAddPlaceholder';
 
 /**
  * @param {Buffer} pdf
@@ -160,66 +135,6 @@ const readPdf = (pdf) => {
     };
 };
 
-/**
- * Adds the objects that are needed for Adobe.PPKLite to read the signature.
- * Also includes a placeholder for the actual signature.
- * Returns an Object with all the added PDFReferences.
- * @param {PDFDocument} pdf
- * @param {string} reason
- * @returns {object}
- */
-export const addSignaturePlaceholder = ({
-    pdf,
-    reason,
-    signatureLength = DEFAULT_SIGNATURE_LENGTH,
-}) => {
-    /* eslint-disable no-underscore-dangle,no-param-reassign */
-    // Generate the signature placeholder
-    const signature = pdf.ref({
-        Type: 'Sig',
-        Filter: 'Adobe.PPKLite',
-        SubFilter: 'adbe.pkcs7.detached',
-        ByteRange: [
-            0,
-            DEFAULT_BYTE_RANGE_PLACEHOLDER,
-            DEFAULT_BYTE_RANGE_PLACEHOLDER,
-            DEFAULT_BYTE_RANGE_PLACEHOLDER,
-        ],
-        Contents: Buffer.from(String.fromCharCode(0).repeat(signatureLength)),
-        Reason: new String(reason), // eslint-disable-line no-new-wrappers
-        M: new Date(),
-    });
-
-    // Generate signature annotation widget
-    const widget = pdf.ref({
-        Type: 'Annot',
-        Subtype: 'Widget',
-        FT: 'Sig',
-        Rect: [0, 0, 0, 0],
-        V: signature,
-        T: new String('Signature1'), // eslint-disable-line no-new-wrappers
-        F: 4,
-        P: pdf.page.dictionary, // eslint-disable-line no-underscore-dangle
-    });
-    // Include the widget in a page
-    pdf.page.dictionary.data.Annots = [widget];
-
-    // Create a form (with the widget) and link in the _root
-    const form = pdf.ref({
-        Type: 'AcroForm',
-        SigFlags: 3,
-        Fields: [widget],
-    });
-    pdf._root.data.AcroForm = form;
-
-    return {
-        signature,
-        form,
-        widget,
-    };
-    /* eslint-enable no-underscore-dangle,no-param-reassign */
-};
-
 const createBufferRootWithAcroform = (pdf, info, form) => {
     const rootIndex = getIndexFromRef(info.xref, info.rootRef);
 
@@ -302,7 +217,7 @@ const getPageRef = (pdf, info) => {
 /**
  * @param {Buffer} pdf
  */
-export const plainAdd = (pdfBuffer, {reason, signatureLength = DEFAULT_SIGNATURE_LENGTH}) => {
+const plainAddPlaceholder = (pdfBuffer, {reason, signatureLength = DEFAULT_SIGNATURE_LENGTH}) => {
     let pdf = removeTrailingNewLine(pdfBuffer);
     const info = readPdf(pdf);
     const pageRef = getPageRef(pdf, info);
@@ -339,11 +254,10 @@ export const plainAdd = (pdfBuffer, {reason, signatureLength = DEFAULT_SIGNATURE
         },
     };
 
-    fs.createWriteStream('./_before.pdf').end(pdf);
     const {
         form,
         widget,
-    } = addSignaturePlaceholder({
+    } = pdfkitAddPlaceholder({
         pdf: pdfKitMock,
         reason,
         signatureLength,
@@ -370,49 +284,7 @@ export const plainAdd = (pdfBuffer, {reason, signatureLength = DEFAULT_SIGNATURE
         createBufferTrailer(pdf, info, addedReferences),
     ]);
 
-    fs.createWriteStream('./_after.pdf').end(pdf);
-
     return pdf;
 };
 
-export const extractSignature = (pdf) => {
-    const byteRangePos = pdf.indexOf('/ByteRange [');
-    if (byteRangePos === -1) {
-        throw new SignPdfError(
-            'Failed to locate ByteRange.',
-            SignPdfError.TYPE_PARSE,
-        );
-    }
-
-    const byteRangeEnd = pdf.indexOf(']', byteRangePos);
-    if (byteRangeEnd === -1) {
-        throw new SignPdfError(
-            'Failed to locate the end of the ByteRange.',
-            SignPdfError.TYPE_PARSE,
-        );
-    }
-
-    const byteRange = pdf.slice(byteRangePos, byteRangeEnd + 1).toString();
-    const matches = (/\/ByteRange \[(\d+) +(\d+) +(\d+) +(\d+)\]/).exec(byteRange);
-
-    const signedData = Buffer.concat([
-        pdf.slice(
-            parseInt(matches[1]),
-            parseInt(matches[1]) + parseInt(matches[2]),
-        ),
-        pdf.slice(
-            parseInt(matches[3]),
-            parseInt(matches[3]) + parseInt(matches[4]),
-        ),
-    ]);
-
-    let signatureHex = pdf.slice(
-        parseInt(matches[1]) + parseInt(matches[2]) + 1,
-        parseInt(matches[3]) - 1,
-    ).toString('binary');
-    signatureHex = signatureHex.replace(/(?:00)*$/, '');
-
-    const signature = Buffer.from(signatureHex, 'hex').toString('binary');
-
-    return {ByteRange: matches.slice(1, 5).map(Number), signature, signedData};
-};
+export default plainAddPlaceholder;
